env valid wip
