+++ src/server/db.ts
@@ -3,18 +3,15 @@ import dotenv from "dotenv";
 import { Kysely, PostgresDialect } from "kysely";
 import type { DB } from "kysely-codegen/dist/db";
 import { Pool } from "pg";
+import { env } from "~/utils/envValidator.js";
 
 dotenv.config();
 
 export const pool = new Pool({
-  //TODO change these two to .env
-  host:
-    process.env.NODE_ENV === "production"
-      ? process.env.PG_SERVICE
-      : "127.0.0.1",
-  database: process.env.NODE_ENV === "production" ? "stories" : "stories_dev",
-  password: process.env.PSQL_PASSWORD,
-  user: process.env.PSQL_USERNAME,
+  host: env.PSQL_HOST,     
+  database: env.PSQL_DATABASE,
+  password: env.PSQL_PASSWORD,
+  user: env.PSQL_USERNAME,
   port: 5432,
 });
